requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
catboost
xgboost
Flask
python-binance
dotenv
beautifulsoup4

#-e .